pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rnaseq-analysis"
version = "0.1.0"
description = "Configurable pairwise RNA-seq differential analysis pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "matplotlib>=3.7",
    "statsmodels>=0.14",
    "PyYAML>=6.0",
    "openpyxl>=3.1",
    "pydeseq2>=0.4",
]

[project.optional-dependencies]
gsea = ["gseapy>=1.0"]
preprocessing = []  # fastp, hisat2, samtools, htseq-count are system tools

[project.scripts]
rnaseq-analysis = "scripts.__main__:main"

[tool.setuptools.packages.find]
include = ["scripts", "scripts.*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
